Extract program factories from slice_program_factory.cpp
- Extract slice_rm_multi_core function and helpers to separate files
- Extract slice_rm_strided_single_core_n_dims function to separate files
- Extract slice_rm_multi_core_sharded function and helpers to separate files
- Extract slice_tile_multi_core function and helpers to separate files
- Update slice_program_factory.cpp to only contain dispatcher function
- Update slice_program_factory.hpp with declarations for extracted functions
