+++ ttnn/cpp/ttnn/operations/data_movement/slice/device/slice_program_factory.hpp
@@ -7,6 +7,22 @@
 
 namespace ttnn::operations::data_movement::detail {
 
+tt::tt_metal::operation::ProgramWithCallbacks slice_rm_multi_core(
+    const Tensor& a, Tensor& output, const ttnn::Shape& output_tensor_start, const ttnn::Shape& output_tensor_end);
+
+tt::tt_metal::operation::ProgramWithCallbacks slice_rm_strided_single_core_n_dims(
+    const Tensor& a,
+    Tensor& output,
+    const ttnn::Shape& output_tensor_start,
+    const ttnn::Shape& output_tensor_end,
+    const ttnn::Shape& step);
+
+tt::tt_metal::operation::ProgramWithCallbacks slice_rm_multi_core_sharded(
+    const Tensor& a, Tensor& output, const ttnn::Shape& output_tensor_start, const ttnn::Shape& output_tensor_end);
+
+tt::tt_metal::operation::ProgramWithCallbacks slice_tile_multi_core(
+    const Tensor& a, Tensor& output, const ttnn::Shape& output_tensor_start, const ttnn::Shape& output_tensor_end);
+
 tt::tt_metal::operation::ProgramWithCallbacks slice_multi_core(
     const Tensor& a,
     Tensor& output,
