+++ ttnn/cpp/ttnn/operations/data_movement/slice/device/slice_rm_multi_core_sharded.cpp
@@ -1,0 +1,319 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "slice_rm_multi_core_sharded.hpp"
+#include "optional"
+#include <tt-metalium/constants.hpp>
+#include <tt-metalium/host_api.hpp>
+
+#include "slice_op.hpp"
+using namespace tt::constants;
+using namespace tt::tt_metal;
+
+namespace ttnn::operations::data_movement::detail {
+
+inline std::vector<std::vector<uint32_t>> group_contiguous_values(std::vector<uint32_t>& values) {
+    std::vector<std::vector<uint32_t>> chunks;
+    if (values.empty()) {
+        return chunks;
+    }
+
+    // Initialize the first chunk
+    std::vector<uint32_t> current_chunk;
+    current_chunk.push_back(values[0]);
+
+    for (size_t i = 1; i < values.size(); ++i) {
+        if (values[i] == values[i - 1] + 1) {
+            current_chunk.push_back(values[i]);
+        } else {
+            chunks.push_back(current_chunk);
+            current_chunk.clear();
+            current_chunk.push_back(values[i]);
+        }
+    }
+    // Add the last chunk
+    chunks.push_back(current_chunk);
+    return chunks;
+}
+
+inline std::vector<std::pair<std::vector<uint32_t>, std::vector<uint32_t>>> get_slice_runtime_args_rm_sharded(
+    const Tensor& input_tensor,
+    Tensor& output_tensor,
+    const ttnn::Shape& output_tensor_start,
+    uint32_t num_cores_unpadded,
+    bool row_major,
+    uint32_t num_cores_x_unpadded,
+    uint32_t num_cores_y_unpadded,
+    uint32_t shard_height_unpadded,
+    uint32_t shard_height_padded,
+    uint32_t num_cores_x_padded,
+    uint32_t num_cores_y_padded) {
+    tt::tt_metal::IDevice* device = input_tensor.device();
+
+    auto output_buffer = output_tensor.buffer();
+    auto input_shape = input_tensor.padded_shape();
+    auto output_shape = output_tensor.padded_shape();
+
+    uint32_t padded_row_size_bytes = input_shape[-1] * input_tensor.element_size();
+    uint32_t unpadded_row_size_bytes = output_shape[-1] * input_tensor.element_size();
+
+    std::uint32_t num_dims = static_cast<std::uint32_t>(input_shape.rank());
+    std::vector<uint32_t> num_unpadded_sticks_per_dim(num_dims);
+    std::vector<uint32_t> num_padded_sticks_per_dim(num_dims);
+    std::vector<uint32_t> id_per_dim(num_dims);
+
+    std::vector<uint32_t> accumulated_total_per_dim(num_dims);
+
+    // TODO: Remove first element of these arrays and update kernel accordingly
+    // This currently just matches tile version where we iterate over the row as well
+    num_unpadded_sticks_per_dim[0] = 1;
+    num_padded_sticks_per_dim[0] = 0;
+    accumulated_total_per_dim[0] = 1;
+
+    for (int32_t i = 1; i < num_dims; i++) {
+        uint32_t num_unpadded_dim = output_shape[-(i + 1)];
+        uint32_t num_total_dim = input_shape[-(i + 1)];
+        uint32_t num_padded_dim = (num_total_dim - num_unpadded_dim) * accumulated_total_per_dim[i - 1];
+        num_unpadded_sticks_per_dim[i] = num_unpadded_dim;
+        num_padded_sticks_per_dim[i] = num_padded_dim;
+        accumulated_total_per_dim[i] = num_total_dim * accumulated_total_per_dim[i - 1];
+    }
+
+    uint32_t unpadded_row_size_bytes_offset = tt::round_up(unpadded_row_size_bytes, TILE_WIDTH / 2);
+
+    std::vector<std::pair<std::vector<uint32_t>, std::vector<uint32_t>>> ret_val(num_cores_unpadded);
+
+    uint32_t start_offset = ttnn::operations::data_movement::get_rm_start_offset(input_tensor, output_tensor_start);
+    for (uint32_t i = 0, num_sticks_written = 0; i < num_cores_unpadded; i++) {
+        CoreCoord core;
+        if (row_major) {
+            core = {i % num_cores_x_unpadded, i / num_cores_x_unpadded};
+        } else {
+            core = {i / num_cores_y_unpadded, i % num_cores_y_unpadded};
+        }
+        uint32_t num_sticks_per_core_unpadded = shard_height_unpadded;
+        uint32_t num_sticks_per_core_padded = shard_height_padded;
+
+        // figure out the start read stick id for each core, and the start id for each dim
+        id_per_dim[0] = num_sticks_written % num_unpadded_sticks_per_dim[0];
+        uint32_t unpadded_written = num_sticks_written / num_unpadded_sticks_per_dim[0];
+        uint32_t start_id = id_per_dim[0] + start_offset;
+
+        for (uint32_t j = 1; j < num_dims; j++) {
+            id_per_dim[j] = unpadded_written % num_unpadded_sticks_per_dim[j];
+            unpadded_written = unpadded_written / num_unpadded_sticks_per_dim[j];
+            start_id += id_per_dim[j] * accumulated_total_per_dim[j - 1];
+        }
+
+        num_sticks_written += num_sticks_per_core_unpadded;
+
+        // stores all sticks id for a core
+        std::vector<uint32_t> stick_ids_per_core;
+        uint32_t src_stick_id = start_id;
+        for (uint32_t i = 0; i < num_sticks_per_core_unpadded; ++i) {
+            stick_ids_per_core.push_back(src_stick_id);
+            src_stick_id++;
+            for (uint32_t j = 0; j < num_dims; j++) {
+                id_per_dim[j]++;
+                if (id_per_dim[j] == num_unpadded_sticks_per_dim[j]) {
+                    id_per_dim[j] = 0;
+                    src_stick_id += num_padded_sticks_per_dim[j];
+                } else {
+                    break;
+                }
+            }
+        }
+
+        // figure out the stick id in a shard, and the core id for the stick.
+        std::map<std::pair<uint32_t, uint32_t>, std::vector<uint32_t>> core_stick_map;
+        for (uint32_t i = 0; i < num_sticks_per_core_unpadded; ++i) {
+            uint32_t stick_id = stick_ids_per_core[i];
+            uint32_t shard_id = stick_id / num_sticks_per_core_padded;
+            uint32_t stick_id_in_shard = stick_id - (shard_id * num_sticks_per_core_padded);
+
+            uint32_t shard_grid_inner_dim = row_major ? num_cores_x_padded : num_cores_y_padded;
+            uint32_t shard_grid_outer_dim_id = shard_id / shard_grid_inner_dim;
+            uint32_t shard_grid_inner_dim_id = shard_id - (shard_grid_outer_dim_id * shard_grid_inner_dim);
+
+            uint32_t worker_y_logical = row_major ? shard_grid_outer_dim_id : shard_grid_inner_dim_id;
+            uint32_t worker_x_logical = row_major ? shard_grid_inner_dim_id : shard_grid_outer_dim_id;
+
+            if (worker_x_logical < num_cores_x_padded and worker_y_logical < num_cores_y_padded) {
+                auto core_physical =
+                    device->worker_core_from_logical_core(CoreCoord{worker_x_logical, worker_y_logical});
+                // save stick id in a shard, and core coord into a map
+                std::pair<uint32_t, uint32_t> xy_pair = row_major ? std::make_pair(core_physical.y, core_physical.x)
+                                                                  : std::make_pair(core_physical.x, core_physical.y);
+                core_stick_map[xy_pair].push_back(stick_id_in_shard);
+            }
+        }
+
+        // reader rt args
+        std::vector<uint32_t> reader_kernel_args;
+        reader_kernel_args.push_back(core_stick_map.size());  // num_cores
+
+        for (const auto& core_stick_pair : core_stick_map) {
+            auto xy_pair = core_stick_pair.first;
+            if (row_major) {
+                reader_kernel_args.push_back(xy_pair.second);  // noc x
+                reader_kernel_args.push_back(xy_pair.first);   // noc y
+            } else {
+                reader_kernel_args.push_back(xy_pair.first);   // noc x
+                reader_kernel_args.push_back(xy_pair.second);  // noc y
+            }
+        }
+
+        // coalesce the sticks into chunks
+        std::vector<std::vector<std::vector<uint32_t>>> stick_chunks_per_core;
+        for (auto core_stick_pair : core_stick_map) {
+            auto stick_chunks = group_contiguous_values(core_stick_pair.second);
+            stick_chunks_per_core.push_back(stick_chunks);
+
+            reader_kernel_args.push_back(stick_chunks.size());  // num_chunks for current core
+        }
+        for (const auto& stick_chunks : stick_chunks_per_core) {
+            for (auto chunk : stick_chunks) {
+                reader_kernel_args.push_back(chunk[0]);      // start id of a chunk
+                reader_kernel_args.push_back(chunk.size());  // length of a chunk
+            }
+        }
+
+        std::vector<uint32_t> writer_kernel_args;
+        ret_val[i] = {reader_kernel_args, writer_kernel_args};
+    }
+
+    return ret_val;
+}
+
+operation::ProgramWithCallbacks slice_rm_multi_core_sharded(
+    const Tensor& a, Tensor& output, const ttnn::Shape& output_tensor_start, const ttnn::Shape& output_tensor_end) {
+    const ttnn::Shape output_shape = output.padded_shape();
+
+    tt::tt_metal::Program program = tt::tt_metal::CreateProgram();
+
+    // This should allocate a DRAM buffer on the device
+    tt::tt_metal::IDevice* device = a.device();
+
+    uint32_t num_padded_sticks = a.physical_volume() / a.padded_shape()[-1];
+    uint32_t num_unpadded_sticks = output.physical_volume() / output.padded_shape()[-1];
+
+    // stick sizes
+    uint32_t W_padded = a.logical_shape()[-1];
+    uint32_t W_unpadded = output.logical_shape()[-1];
+    auto stick_size_padded = W_padded * a.element_size();
+    auto stick_size_unpadded = W_unpadded * output.element_size();
+
+    // input shard spec
+    auto shard_spec_padded = a.shard_spec().value();
+    uint32_t shard_height_padded = shard_spec_padded.shape[0];
+    uint32_t shard_width_padded = shard_spec_padded.shape[1];
+
+    auto& all_cores_padded = shard_spec_padded.grid;
+    uint32_t num_cores_padded = shard_spec_padded.num_cores();
+    auto bbox_padded = shard_spec_padded.grid.bounding_box();
+    CoreCoord grid_size_padded = {bbox_padded.end_coord.x + 1, bbox_padded.end_coord.y + 1};
+    uint32_t num_cores_x_padded = grid_size_padded.x;
+    uint32_t num_cores_y_padded = grid_size_padded.y;
+
+    log_debug(tt::LogOp, "num_padded_sticks: {}", num_padded_sticks);
+    log_debug(tt::LogOp, "shard_height_padded: {}", shard_height_padded);
+    log_debug(tt::LogOp, "all_cores_padded: {}", all_cores_padded);
+    log_debug(tt::LogOp, "num_cores_padded: {}", num_cores_padded);
+
+    // output shard spec
+    auto shard_spec_unpadded = output.shard_spec().value();
+    uint32_t shard_height_unpadded = shard_spec_unpadded.shape[0];
+    uint32_t shard_width_unpadded = shard_spec_unpadded.shape[1];
+    bool row_major = shard_spec_unpadded.orientation == ShardOrientation::ROW_MAJOR;
+
+    auto& all_cores_unpadded = shard_spec_unpadded.grid;
+    uint32_t num_cores_unpadded = shard_spec_unpadded.num_cores();
+    auto bbox_unpadded = shard_spec_unpadded.grid.bounding_box();
+    CoreCoord grid_size_unpadded = {bbox_unpadded.end_coord.x + 1, bbox_unpadded.end_coord.y + 1};
+    uint32_t num_cores_x_unpadded = grid_size_unpadded.x;
+    uint32_t num_cores_y_unpadded = grid_size_unpadded.y;
+
+    log_debug(tt::LogOp, "num_unpadded_sticks: {}", num_unpadded_sticks);
+    log_debug(tt::LogOp, "shard_height_unpadded: {}", shard_height_unpadded);
+    log_debug(tt::LogOp, "all_cores_unpadded: {}", all_cores_unpadded);
+    log_debug(tt::LogOp, "num_cores_unpadded: {}", num_cores_unpadded);
+
+    tt::tt_metal::Buffer* src0_buffer = a.buffer();
+
+    tt::DataFormat cb_data_format = tt::tt_metal::datatype_to_dataformat_converter(a.dtype());
+    tt::DataFormat dst_cb_data_format = tt::tt_metal::datatype_to_dataformat_converter(output.dtype());
+
+    tt::tt_metal::Buffer* dst_buffer = output.buffer();
+    TT_ASSERT(dst_buffer != nullptr, "Output buffer should be allocated on device!");
+
+    auto compute_with_storage_grid_size = device->compute_with_storage_grid_size();
+    uint32_t num_cores_x = compute_with_storage_grid_size.x;
+    uint32_t num_cores_y = compute_with_storage_grid_size.y;
+    CoreRange total_cores({0, 0}, {num_cores_x - 1, num_cores_y - 1});
+
+    uint32_t src0_cb_index = 0;
+    tt::tt_metal::CircularBufferConfig cb_src0_config =
+        tt::tt_metal::CircularBufferConfig(shard_height_padded * stick_size_padded, {{src0_cb_index, cb_data_format}})
+            .set_page_size(src0_cb_index, stick_size_padded)
+            .set_globally_allocated_address(*a.buffer());
+    auto cb_src0 = tt::tt_metal::CreateCircularBuffer(program, total_cores, cb_src0_config);
+
+    uint32_t output_cb_index = tt::CBIndex::c_16;
+    tt::tt_metal::CircularBufferConfig cb_output_config =
+        tt::tt_metal::CircularBufferConfig(
+            shard_height_unpadded * stick_size_unpadded, {{output_cb_index, dst_cb_data_format}})
+            .set_page_size(output_cb_index, stick_size_unpadded)
+            .set_globally_allocated_address(*output.buffer());
+    auto cb_output = tt::tt_metal::CreateCircularBuffer(program, total_cores, cb_output_config);
+
+    std::vector<uint32_t> reader_ct_args = {
+        (std::uint32_t)stick_size_padded, (std::uint32_t)stick_size_unpadded, (std::uint32_t)shard_height_unpadded};
+
+    tt::tt_metal::KernelHandle unary_reader_kernel_id = tt::tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/operations/data_movement/slice/device/kernels/dataflow/"
+        "slice_reader_unary_unpad_dims_rm_sharded.cpp",
+        all_cores_unpadded,
+        tt::tt_metal::ReaderDataMovementConfig(reader_ct_args));
+
+    auto all_runtime_args = get_slice_runtime_args_rm_sharded(
+        a,
+        output,
+        output_tensor_start,
+        num_cores_unpadded,
+        row_major,
+        num_cores_x_unpadded,
+        num_cores_y_unpadded,
+        shard_height_unpadded,
+        shard_height_padded,
+        num_cores_x_padded,
+        num_cores_y_padded);
+
+    for (uint32_t i = 0; i < num_cores_unpadded; i++) {
+        CoreCoord core;
+        if (row_major) {
+            core = {i % num_cores_x_unpadded, i / num_cores_x_unpadded};
+        } else {
+            core = {i / num_cores_y_unpadded, i % num_cores_y_unpadded};
+        }
+        tt::tt_metal::SetRuntimeArgs(program, unary_reader_kernel_id, core, all_runtime_args[i].first);
+    }
+
+    auto override_runtime_args_callback = [cb_src0, cb_output](
+                                              const void* operation,
+                                              Program& program,
+                                              const std::vector<Tensor>& input_tensors,
+                                              const std::vector<std::optional<const Tensor>>&,
+                                              const std::vector<Tensor>& output_tensors) {
+        auto src_buffer_a = input_tensors.at(0).buffer();
+        auto dst_buffer = output_tensors.at(0).buffer();
+
+        UpdateDynamicCircularBufferAddress(program, cb_src0, *src_buffer_a);
+        UpdateDynamicCircularBufferAddress(program, cb_output, *dst_buffer);
+    };
+
+    return {.program = std::move(program), .override_runtime_arguments_callback = override_runtime_args_callback};
+}
+
+}  // namespace ttnn::operations::data_movement::detail
