+++ ttnn/cpp/ttnn/operations/data_movement/slice/device/slice_rm_strided_single_core_n_dims.hpp
@@ -1,0 +1,17 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+#pragma once
+
+#include <tt-metalium/host_api.hpp>
+
+namespace ttnn::operations::data_movement::detail {
+
+tt::tt_metal::operation::ProgramWithCallbacks slice_rm_strided_single_core_n_dims(
+    const Tensor& a,
+    Tensor& output,
+    const ttnn::Shape& output_tensor_start,
+    const ttnn::Shape& output_tensor_end,
+    const ttnn::Shape& step);
+
+}  // namespace ttnn::operations::data_movement::detail
